feat: implement React Context state management for GoPCA Desktop (#505)
- Introduce FileDataContext, PCAContext, VisualizationContext, UIContext, GoCSVContext
  wrapping existing custom hooks so any component can access shared state without prop drilling
- Extract DataLoadSection, PCAConfigSection, ResultsSection JSX sub-components from App.tsx;
  App.tsx is now a clean provider stack + AppContent orchestration shell
- Add Vitest + @testing-library/react test infrastructure (jsdom, Wails mocks, ui-components mock)
- Write 21 tests across all 5 contexts covering initial state, setter/handler types, and
  "throws outside provider" guards

Fixes #505
