+++ cmd/gopca-desktop/frontend/vite.config.ts
@@ -4,10 +4,25 @@
 // The author respectfully requests that it not be used for
 // military, warfare, or surveillance applications.
 
+/// <reference types="vitest" />
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()]
+  plugins: [react()],
+  test: {
+    environment: 'jsdom',
+    globals: true,
+    setupFiles: ['./src/test/setup.ts'],
+    include: ['src/**/*.{test,spec}.{ts,tsx}'],
+    alias: {
+      // Redirect Wails bindings and ui-components to mocks in test environment
+      '../wailsjs/go/main/App': '/Users/runema/Development/gopca/cmd/gopca-desktop/frontend/src/__mocks__/wailsjs/go/main/App.ts',
+      '../../wailsjs/go/main/App': '/Users/runema/Development/gopca/cmd/gopca-desktop/frontend/src/__mocks__/wailsjs/go/main/App.ts',
+      '../wailsjs/runtime/runtime': '/Users/runema/Development/gopca/cmd/gopca-desktop/frontend/src/__mocks__/wailsjs/runtime/runtime.ts',
+      '../../wailsjs/runtime/runtime': '/Users/runema/Development/gopca/cmd/gopca-desktop/frontend/src/__mocks__/wailsjs/runtime/runtime.ts',
+      '@gopca/ui-components': '/Users/runema/Development/gopca/cmd/gopca-desktop/frontend/src/__mocks__/@gopca/ui-components.tsx',
+    },
+  },
 });
